Add trailing slashes to hrefs
This change is necessary since static-adapter converts files like projects.svelte to project/index.html and this causes the page to be served with trailing slash.

Google Search Console looks for the links in the pages and the lack of trailing slash causes a page to be not indexed by Google.